lab11_1_fianl

+++ lab11_1.cpp
@@ -1,6 +1,6 @@
 #include<iostream>
 using namespace std;
-int fibonacci(int);
+long long int fibonacci(int);
 
 int main(){
     int x;
@@ -9,7 +9,7 @@ int main(){
     return 0; 
 }
 
-int fibonacci(int x)
+long long int fibonacci(int x)
 {
 	if(x==0)
 		{return 0;}
